changes functions placement in provisioning and connector files. fixes interpolation for Header for OAuth. Comments out into examples for provisioning examples
