+++ pkg/util/utils.go
@@ -19,7 +19,7 @@ const (
 	LegacyPem        = "legacy-pem"
 	HeaderTpplApikey = "tppl-api-key"
 	OauthTokenType   = "Bearer"
-	HeaderOauthToken = OauthTokenType + " %s"
+	HeaderOauthToken = OauthTokenType + " "
 	DefaultTimeout   = 180 // seconds
 )
