+++ pkg/venafi/cloud/provisioning.go
@@ -3,16 +3,14 @@ package cloud
 import (
 	"encoding/json"
 	"fmt"
-	"golang.org/x/net/context"
-	"log"
 	"time"
 
-	"github.com/google/uuid"
+	"github.com/Khan/genqlient/graphql"
 
 	"github.com/Venafi/vcert/v5/internal/datasource/cloudkeystores"
 	"github.com/Venafi/vcert/v5/pkg/endpoint"
-	"github.com/Venafi/vcert/v5/pkg/util"
-	"github.com/Venafi/vcert/v5/pkg/websocket"
+	"github.com/Venafi/vcert/v5/pkg/framework"
+	"net/http"
 )
 
 type CloudKeystoreProvisioningResult struct {
@@ -118,104 +116,6 @@ func (cpgo CloudProvisioningGCPOptions) GetType() string {
 	return "GCM"
 }
 
-func (c *Connector) ProvisionCertificate(req *endpoint.ProvisioningRequest, options *endpoint.ProvisioningOptions) (provisioningMetadata endpoint.ProvisioningMetadata, err error) {
-	log.Printf("Starting Provisioning Flow")
-
-	if req == nil {
-		return nil, fmt.Errorf("missing Provisioning Request")
-	}
-
-	reqData := *req
-
-	if reqData.Timeout == 0 {
-		reqData.Timeout = util.DefaultTimeout * time.Second
-	}
-
-	if reqData.KeystoreID == nil {
-		return nil, fmt.Errorf("keystore ID was not provided for provisioning")
-	}
-	keystoreIDString := *(reqData.KeystoreID)
-	log.Printf("Keystore ID for provisioning: %s", keystoreIDString)
-
-	if reqData.CertificateID == nil {
-		if reqData.PickupID == nil {
-			return nil, fmt.Errorf("no Certificate ID or Pickup ID were provided for provisioning")
-		}
-		log.Printf("Certificate ID was not provided in request. Fetching it by using Pickup ID %s", *(reqData.PickupID))
-		certID, err := c.getCertIDFromPickupID(*(reqData.PickupID), reqData.Timeout)
-		if err != nil {
-			return nil, err
-		}
-		reqData.CertificateID = certID
-	}
-	certificateIDString := *(reqData.CertificateID)
-	log.Printf("Certificate ID for provisioning: %s", certificateIDString)
-
-	// Is certificate generated by VCP?
-	log.Printf("Validating if certificate is generated by VCP")
-	err = c.validateIfCertIsVCPGeneratedByID(*(reqData.CertificateID))
-	if err != nil {
-		return nil, err
-	}
-	log.Println("Certificate is valid for provisioning (VCP generated)")
-
-	// setting options for provisioning
-	log.Println("setting provisioning options")
-	provisioningOptions, err := setProvisioningOptions(options)
-	if err != nil {
-		return nil, err
-	}
-	log.Println("provisioning options successfully set")
-
-	graphqlClient := c.getGraphqlClient()
-
-	ctx := context.Background()
-
-	// Getting Keystore to find type
-	log.Printf("fetching keystore information for KeystoreID: %s", keystoreIDString)
-	data, err := cloudkeystores.GetCloudKeystoresByKeystoreId(ctx, graphqlClient, req.KeystoreID)
-	if err != nil {
-		return nil, err
-	}
-
-	if len(data.CloudKeystores.Nodes) != 1 {
-		return nil, fmt.Errorf("could not find keystore with ID: %s", keystoreIDString)
-	}
-
-	// grabbing found keystore for later
-	cloudKeystore := data.CloudKeystores.Nodes[0]
-	log.Printf("successfully fetched keystore information for KeystoreID: %s", keystoreIDString)
-
-	wsClientID := uuid.New().String()
-
-	wsConn, err := websocket.Subscribe(c.apiKey, c.baseURL, wsClientID)
-	if err != nil {
-		return nil, err
-	}
-
-	log.Printf("Provisioning Certificate ID %s for Keystore %s", certificateIDString, keystoreIDString)
-	_, err = cloudkeystores.ProvisionCertificate(ctx, graphqlClient, certificateIDString, keystoreIDString, wsClientID, provisioningOptions)
-	if err != nil {
-		return nil, err
-	}
-
-	ar, err := websocket.ReadResponse(wsConn)
-	if err != nil {
-		return nil, err
-	}
-
-	// parsing metadata from websocket response
-	log.Printf("Getting Cloud Metadata of Certificate ID %s and Keystore ID: %s", certificateIDString, keystoreIDString)
-	cloudMetadata, err := getCloudMetadataFromWebsocketResponse(ar.Data.Result, cloudKeystore.Type, keystoreIDString)
-	if err != nil {
-		return nil, err
-	}
-	log.Printf("Successfully got Cloud Metadata for Certificate ID %s and Keystore ID: %s", certificateIDString, keystoreIDString)
-
-	log.Printf("Successfully finished Provisioning Flow for Certificate ID %s and Keystore ID %s", certificateIDString, keystoreIDString)
-	return cloudMetadata, nil
-}
-
 func setProvisioningOptions(options *endpoint.ProvisioningOptions) (*cloudkeystores.CertificateProvisioningOptionsInput, error) {
 	var cloudOptions *cloudkeystores.CertificateProvisioningOptionsInput
 	dataOptions, err := json.Marshal(options)
@@ -251,4 +151,67 @@ func setProvisioningOptions(options *endpoint.ProvisioningOptions) (*cloudkeysto
 		GcpOptions:   graphqlGCPOptions,
 	}
 	return cloudOptions, nil
+}
+
+func (c *Connector) validateIfCertIsVCPGeneratedByID(certificateId string) error {
+	cert, err := c.getCertificates(certificateId)
+	if err != nil {
+		return fmt.Errorf("error trying to get certificate details for cert with ID: %s, error: %s", certificateId, err.Error())
+	}
+	if cert.DekHash == "" {
+		return fmt.Errorf("error trying to provisioning certificate with ID: %s. Provided certificate is not VCP generated", certificateId)
+	}
+	return nil
+}
+
+func (c *Connector) getGraphqlClient() graphql.Client {
+	graphqlURL := c.getURL(urlGraphql)
+
+	// We provide every type of auth here.
+	// The logic to decide which auth is inside struct's function: RoundTrip
+	httpclient := &http.Client{
+		Transport: &framework.AuthedTransportApi{
+			ApiKey:      c.apiKey,
+			AccessToken: c.accessToken,
+			Wrapped:     http.DefaultTransport,
+		},
+		Timeout: 30 * time.Second,
+	}
+
+	client := graphql.NewClient(graphqlURL, httpclient)
+	return client
+}
+
+func getCloudMetadataFromWebsocketResponse(respMap interface{}, keystoreType cloudkeystores.CloudKeystoreType, keystoreId string) (*CloudProvisioningMetadata, error) {
+
+	val := CloudKeystoreProvisioningResult{}
+	valJs, err := json.Marshal(respMap)
+	if err != nil {
+		return nil, fmt.Errorf("unable to encode response data! Error: %s", err.Error())
+	}
+	err = json.Unmarshal(valJs, &val)
+	if err != nil {
+		return nil, fmt.Errorf("unable to parse response data! Error: %s", err.Error())
+	}
+	if val.Error != nil {
+		return nil, fmt.Errorf("unable to provision certificate! Error: %s", val.Error)
+	}
+
+	if val.CloudProviderCertificateID == "" {
+		return nil, fmt.Errorf("provisioning is not successful, certificate ID from response is empty")
+	}
+
+	cloudMetadata := &CloudProvisioningMetadata{}
+	switch keystoreType {
+	case cloudkeystores.CloudKeystoreTypeAcm:
+		cloudMetadata.awsMetadata.result = val
+	case cloudkeystores.CloudKeystoreTypeAkv:
+		cloudMetadata.azureMetadata.result = val
+	case cloudkeystores.CloudKeystoreTypeGcm:
+		cloudMetadata.gcpMetadata.result = val
+	default:
+		err = fmt.Errorf("Unknown ConnectorType %v found for keystore with ID: %s", keystoreType, keystoreId)
+		return nil, err
+	}
+	return cloudMetadata, err
 }
